+++ CHANGELOG.md
@@ -12,10 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Access on hidden users for `superadmin`
 - Redirect to same user page where "click" happened
 - `nowrap` CSS class to table values
+- [`FileExtension`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Twig/FileExtension.php)
+- Twig filter `file_size` in [`FileExtension`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Twig/FileExtension.php)
 
 ### Changed
 - `$settings` to `$this->settings` in [`FileController`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Controller/FileController.php)
 - Margins in forms
+- Templates with new `file_size` twig filter
+- [`File::$size`](https://github.com/InsanityMeetsHH/file-sharing/blob/master/src/Entity/File.php#L50) from `integer` to `string`
 
 ## [2.1.0] - 2018-10-10
 ### Added